Refactor POS assets and convert JS to ES module format.
Updated POS asset references in `__manifest__.py` and renamed `PaymentScreenStatus.js` to `PaymentScreenStatus.esm.js`, converting it to ES module syntax. Minor improvements include CSS adjustments, better maintainability for custom patches, and enhanced markup for maintainers' information.
